feat(hyprland): Added split-monitor-workspaces plugin to hyprland
This forces us to use hyprland from upstream via an input as there is
no ABI defined and the plugin best be compiled with the matching
hyprland version

The plugin just batches up 10 workspaces and assigns them to each
display. Thus 1-10 is display 1, 11-20 is display 2, and so on. We
rename the workspaces in waybar via icon-formatting to 1-10 repeatedly.